refactor: 번역 UI 셸을 클라이언트 콘텐츠 컴포넌트로 분리
